3D fluid wip

+++ include/pr/view3d-12/model/model_generator.h
@@ -355,7 +355,7 @@ namespace pr::rdr12
 			{
 				assert(vcount >= 0 && icount >= 0 && ncount >= 0 && idx_stride >= 1);
 				m_vcont.resize(vcount);
-				m_icont.resize(icount * idx_stride);
+				m_icont.resize(size_t(icount) * idx_stride);
 				m_ncont.resize(ncount);
 				m_icont.m_stride = idx_stride;
 				m_in_use = true;
